manual-scaler-controller/code/requirements.txt
google-cloud-monitoring>=2.0.0
google-cloud-logging>=1.15.1
google-api-python-client>=1.12.3
google-cloud-datastore>=1.15.3
google-cloud-core>=1.4.3
oauth2client>=4.1.3
google-cloud-scheduler>=2.0.0
flask